Removal of gratuitous casts
git-svn-id: file:///tmp/orig/avr-evtd/trunk@151 a2d00c94-3d26-0410-a74b-ce61611494ea

+++ avr-evtd.c
@@ -654,7 +654,7 @@ static void avr_evtd_main(void)
 			if ((tt_Power_Press + SP_MONITOR_TIME) < tt_TimeNow
 			    && FirstTimeFlag > 1) {
 				/* Inform the EventScript */
-				execute_command(FIVE_SHUTDOWN, (int) ((float) ShutdownTimer / 60.0f));
+				execute_command(FIVE_SHUTDOWN, (int) (ShutdownTimer/60.0);
 				FirstTimeFlag = 1;
 				tt_Power_Press = 0;
 			}
